+++ sidebars.js
@@ -858,15 +858,14 @@ const sidebars = {
         id: "release-notes",
       },
       items: [
+        "enterprise-v3.1",
         "enterprise-v3.0",
         "enterprise-v2.8",
         "enterprise-v2.7",
-        "enterprise-v2.6",
-        "enterprise-v2.5",
         {
           type: "category",
           label: "Older 2.x Releases",
-          items: ["enterprise-v2.4", "enterprise-v2.3"],
+          items: ["enterprise-v2.6", "enterprise-v2.5", "enterprise-v2.4", "enterprise-v2.3"],
         },
         // "v2-release-notes",
         {